requirements.txt
streamlit
scikit-learn
pandas
joblib
python-dotenv
supabase
